images: Move functions into parse_helpers.go to resemble ocicrypt
ocicrypt has a parse_helpers.go that has similar functions as those in
crypt_utils.go. To resemble ocicrypt's parse_helpers.go move functions
out of crypt_utils.go into our own parse_helpers.go. Ultimately we should
be using ocicrypt's parse_helpers.go and get rid of ours.
